refactored trade and item routes to seperate files
