+++ views/trade/viewTrade.ejs
@@ -250,7 +250,7 @@
 setInterval( () => {
   if ( userLoggedIn ) $.ajax( {
       // The URL for the request
-      url: "/interface/pollforchanges/",
+      url: "/interface/trade/pollforchanges/",
 
       // Whether this is a POST or GET request
       type: "GET",
