+++ views/user/userHome.ejs
@@ -236,7 +236,7 @@ Panel example:
 setInterval( () => {
   if ( userLoggedIn ) $.ajax( {
       // The URL for the request
-      url: "/interface/pollforchanges/",
+      url: "/interface/trade/pollforchanges/",
 
       // Whether this is a POST or GET request
       type: "GET",
